Replace /upload endpoint with import-collections command
