+++ package.json
@@ -16,6 +16,7 @@
     "delete-collections": "ts-node scripts/deleteCollections.ts",
     "dev": "nodemon --inspect -r ts-node/register src/index.ts",
     "dump-schema": "ts-node scripts/dumpSchema.ts",
+    "import-collections": "ts-node scripts/importCollections.ts",
     "index-for-search": "ts-node scripts/indexForSearch.ts",
     "lint": "tslint 'src/**/*.{ts,tsx}'",
     "pull-database": "./scripts/pull-database.sh",
